backend/requirements.txt
tavily-python>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-community>=0.0.5